Proxy class support added to function calling
Added the option to inform a proxy class to be called instead of the direct call of the tool. It may be useful to in scenarios where the LLM may not respect the data types informed in the JSON schema. In these scenarios we can use a proxy class with the same methods and interface, but the parameters may all be of type string or generic types. In the proxy class we can implement type conversions in order to prevent short dumps caused by incompatible types.